+++ examples/advanced/core/camera/mod.rs
@@ -5,7 +5,7 @@ pub mod camera_replace_proxies;
 pub use camera_replace_proxies::*;
 
 use bevy::prelude::*;
-use bevy_gltf_blueprints::SpawnSet;
+use bevy_gltf_blueprints::GltfBlueprintsSet;
 
 pub struct CameraPlugin;
 impl Plugin for CameraPlugin {
@@ -17,7 +17,7 @@ impl Plugin for CameraPlugin {
 
         .add_systems(Update, 
           (
-            camera_replace_proxies.after(SpawnSet::AfterSpawn),
+            camera_replace_proxies.after(GltfBlueprintsSet::AfterSpawn),
             camera_track,
           )
         )
