+++ examples/advanced/core/mod.rs
@@ -1,4 +1,5 @@
 pub mod camera;
+use bevy_rapier3d::prelude::Velocity;
 pub use camera::*;
 
 pub mod lighting;
@@ -38,13 +39,27 @@ fn spawn_test(
     let range = 5.5;
     let x: f32 = rng.gen_range(-range..range);
     let y: f32 = rng.gen_range(-range..range);
+
+
+    let mut rng = rand::thread_rng();
+    let range = 0.8;
+    let vel_x: f32 = rng.gen_range(-range..range);
+    let vel_y: f32 = rng.gen_range(2.0..2.5);
+    let vel_z: f32 = rng.gen_range(-range..range);
+
+
     let name_index:u64 = rng.gen();
 
     let new_entity = commands.spawn((
       bevy::prelude::Name::from(format!("test{}", name_index)),
       BlueprintName("Health_Pickup".to_string()),
       SpawnHere,
-      TransformBundle::from_transform(Transform::from_xyz(x, 1.0, y))
+      TransformBundle::from_transform(Transform::from_xyz(x, 2.0, y)),
+
+      Velocity {
+        linvel: Vec3::new(vel_x, vel_y, vel_z),
+        angvel: Vec3::new(0.0, 0.0, 0.0),
+      },
     )).id();
     commands.entity(world).add_child(new_entity);
   }
